[BPK-2724] Add theme attribute for flare boxes
