+++ packages/bpk-component-nudger/BpkNudgerStory.scss
@@ -22,7 +22,3 @@
   padding: $bpk-spacing-xs;
   background-color: $bpk-color-blue-500;
 }
-
-.bpk-nudger-configurable {
-  width: 7rem;
-}
