+++ packages/bpk-component-nudger/index.js
@@ -17,8 +17,7 @@
  */
 
 import BpkNudger from './src/BpkNudger';
-import BpkConfigurableNudger from './src/BpkConfigurableNudger';
 import themeAttributes from './src/themeAttributes';
 
 export default BpkNudger;
-export { themeAttributes, BpkConfigurableNudger };
+export { themeAttributes };
