--- packages/bpk-component-nudger/src/BpkConfigurableNudger.js
@@ -1,145 +1,0 @@
-/* eslint-disable react/forbid-prop-types */
-/*
- * Backpack - Skyscanner's Design System
- *
- * Copyright 2016-2019 Skyscanner Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/* @flow strict */
-
-import PropTypes from 'prop-types';
-import React from 'react';
-import BpkButton from 'bpk-component-button';
-import { withButtonAlignment } from 'bpk-component-icon';
-import MinusIcon from 'bpk-component-icon/sm/minus';
-import PlusIcon from 'bpk-component-icon/sm/plus';
-import { cssModules } from 'bpk-react-utils';
-
-import {
-  type CommonProps,
-  COMMON_PROP_TYPES,
-  COMMON_DEFAULT_PROPS,
-} from './common-types';
-import STYLES from './BpkNudger.scss';
-
-const getClassName = cssModules(STYLES);
-
-const AlignedMinusIcon = withButtonAlignment(MinusIcon);
-const AlignedPlusIcon = withButtonAlignment(PlusIcon);
-
-type Props = {
-  ...$Exact<CommonProps>,
-  inputClassName: ?string,
-  formatValue: (a: any) => string,
-  incrementValue: (a: any) => any,
-  decrementValue: (a: any) => any,
-  compareValues: (a: any, b: any) => number,
-};
-
-const BpkConfigurableNudger = (props: Props) => {
-  const {
-    id,
-    min,
-    max,
-    value,
-    onChange,
-    className,
-    inputClassName,
-    increaseButtonLabel,
-    decreaseButtonLabel,
-    buttonType,
-    formatValue,
-    incrementValue,
-    decrementValue,
-    compareValues,
-  } = props;
-  const classNames = [getClassName('bpk-nudger')];
-  if (className) {
-    classNames.push(className);
-  }
-
-  const maxButtonDisabled = compareValues(value, max) >= 0;
-  const minButtonDisabled = compareValues(value, min) <= 0;
-
-  const minusIconClassNames = [getClassName('bpk-nudger__icon')];
-  if (minButtonDisabled) {
-    minusIconClassNames.push(getClassName('bpk-nudger__icon--disabled'));
-  }
-  const plusIconClassNames = [getClassName('bpk-nudger__icon')];
-  if (maxButtonDisabled) {
-    plusIconClassNames.push(getClassName('bpk-nudger__icon--disabled'));
-  }
-
-  const inputStyles = [getClassName('bpk-nudger__input')];
-  if (inputClassName) {
-    inputStyles.push(inputClassName);
-  }
-  if (buttonType === 'outline') {
-    inputStyles.push(getClassName('bpk-nudger__input--outline'));
-  }
-
-  return (
-    <div className={classNames.join(' ')}>
-      <BpkButton
-        secondary={buttonType === 'secondary'}
-        outline={buttonType === 'outline'}
-        iconOnly
-        onClick={() => onChange(decrementValue(value))}
-        disabled={minButtonDisabled}
-        title={decreaseButtonLabel}
-        aria-controls={id}
-        className={getClassName('bpk-nudger__button')}
-      >
-        <AlignedMinusIcon className={minusIconClassNames.join(' ')} />
-      </BpkButton>
-      <input
-        type="text"
-        aria-live="assertive"
-        readOnly
-        value={formatValue(value)}
-        id={id}
-        className={inputStyles.join(' ')}
-      />
-      <BpkButton
-        secondary={buttonType === 'secondary'}
-        outline={buttonType === 'outline'}
-        iconOnly
-        onClick={() => onChange(incrementValue(value))}
-        disabled={maxButtonDisabled}
-        title={increaseButtonLabel}
-        aria-controls={id}
-        className={getClassName('bpk-nudger__button')}
-      >
-        <AlignedPlusIcon className={plusIconClassNames.join(' ')} />
-      </BpkButton>
-    </div>
-  );
-};
-
-BpkConfigurableNudger.propTypes = {
-  ...COMMON_PROP_TYPES,
-  inputClassName: PropTypes.string,
-  formatValue: PropTypes.PropTypes.func.isRequired,
-  incrementValue: PropTypes.func.isRequired,
-  decrementValue: PropTypes.func.isRequired,
-  compareValues: PropTypes.func.isRequired,
-};
-
-BpkConfigurableNudger.defaultProps = {
-  ...COMMON_DEFAULT_PROPS,
-  inputClassName: null,
-};
-
-export default BpkConfigurableNudger;
