+++ packages/bpk-component-nudger/src/BpkNudger.scss
@@ -24,7 +24,10 @@
   }
 
   &__input {
+    $size: $bpk-button-line-height + ($bpk-button-padding-y * 2);
+
     display: inline-block;
+    width: $size;
     padding: $bpk-button-padding-y 0;
     border: none;
     font-size: $bpk-font-size-lg;
@@ -35,12 +38,6 @@
     &--outline {
       background-color: transparent;
       color: $bpk-color-white;
-    }
-
-    &--numeric {
-      $size: $bpk-button-line-height + ($bpk-button-padding-y * 2);
-
-      width: $size;
     }
   }
 }
