+++ packages/bpk-component-nudger/stories.js
@@ -16,8 +16,6 @@
  * limitations under the License.
  */
 
-/* @flow strict */
-
 import React, { Component } from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
@@ -26,15 +24,11 @@ import { cssModules } from 'bpk-react-utils';
 
 import STYLES from './BpkNudgerStory.scss';
 
-import BpkNudger, { BpkConfigurableNudger } from './index';
+import BpkNudger from './index';
 
 const getClassName = cssModules(STYLES);
 
-type State = {
-  value: any,
-};
-
-class NudgerContainer extends Component<{}, State> {
+class NudgerContainer extends Component {
   constructor() {
     super();
 
@@ -59,63 +53,6 @@ class NudgerContainer extends Component<{}, State> {
           onChange={this.handleChange}
           decreaseButtonLabel="Decrease"
           increaseButtonLabel="Increase"
-        />
-      </div>
-    );
-  }
-}
-
-// eslint-disable-next-line react/no-multi-comp
-class ConfigurableNudgerContainer extends Component<{}, State> {
-  constructor() {
-    super();
-
-    this.state = {
-      value: 'premium',
-    };
-  }
-
-  handleChange = value => {
-    this.setState({ value });
-  };
-
-  compareValues = (a, b): number => {
-    const options = ['economy', 'premium', 'business', 'first'];
-    const [aIndex, bIndex] = [options.indexOf(a), options.indexOf(b)];
-    return aIndex - bIndex;
-  };
-
-  incrementValue = (a): string => {
-    const options = ['economy', 'premium', 'business', 'first'];
-    const [aIndex] = [options.indexOf(a) + 1];
-    return options[aIndex];
-  };
-
-  decrementValue = (a): string => {
-    const options = ['economy', 'premium', 'business', 'first'];
-    const [aIndex] = [options.indexOf(a) - 1];
-    return options[aIndex];
-  };
-
-  formatValue = (a): string => a.toString();
-
-  render() {
-    return (
-      <div>
-        <BpkLabel htmlFor="my-nudger">Traveller Class</BpkLabel>
-        <BpkConfigurableNudger
-          id="nudger"
-          min="economy"
-          max="first"
-          value={this.state.value}
-          onChange={this.handleChange}
-          decreaseButtonLabel="Decrease"
-          increaseButtonLabel="Increase"
-          compareValues={this.compareValues}
-          incrementValue={this.incrementValue}
-          decrementValue={this.decrementValue}
-          formatValue={this.formatValue}
-          inputClassName={getClassName('bpk-nudger-configurable')}
         />
       </div>
     );
@@ -170,5 +107,4 @@ storiesOf('bpk-component-nudger', module)
         buttonType="outline"
       />
     </div>
-  ))
-  .add('Configurable nudger', () => <ConfigurableNudgerContainer />);
+  ));
